# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Recruitment API",
        "version": "1.0"
    },
    "paths": {
        "/jobs": {
            "get": {
                "tags": [
                    "Job"
                ],
                "summary": "List of all the available Jobs",
                "operationId": "App\\Http\\Controllers\\JobController::index",
                "responses": {
                    "200": {
                        "description": "All the avilable Jobs."
                    },
                    "default": {
                        "description": "something is bad."
                    }
                }
            },
            "post": {
                "tags": [
                    "Job"
                ],
                "summary": "Job position details",
                "description": "",
                "operationId": "store",
                "requestBody": {
                    "description": "Job information",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Job"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Job registered"
                    },
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/jobs/match/{jobId}": {
            "get": {
                "tags": [
                    "Job"
                ],
                "summary": "List of all the available professionals to this job",
                "operationId": "App\\Http\\Controllers\\JobController::match",
                "parameters": [
                    {
                        "name": "jobId",
                        "in": "path",
                        "description": "Job ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "All the avilable professionals to this job."
                    },
                    "default": {
                        "description": "something is bad."
                    }
                }
            }
        },
        "/registration": {
            "get": {
                "tags": [
                    "Registration"
                ],
                "summary": "List of all the registered candidates",
                "operationId": "App\\Http\\Controllers\\RegistrationController::index",
                "responses": {
                    "200": {
                        "description": "List all candidates."
                    },
                    "default": {
                        "description": "something is bad."
                    }
                }
            },
            "post": {
                "tags": [
                    "Registration"
                ],
                "summary": "Candidate registration",
                "description": "",
                "operationId": "register",
                "requestBody": {
                    "description": "Candidate information",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "person": {
                                        "$ref": "#/components/schemas/Person"
                                    },
                                    "background": {
                                        "items": {
                                            "$ref": "#/components/schemas/Background"
                                        }
                                    },
                                    "work_experience": {
                                        "items": {
                                            "$ref": "#/components/schemas/WorkExperience"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Candidate registered"
                    },
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "List of all the work categories",
                "operationId": "App\\Http\\Controllers\\WorkCategoryController::index",
                "responses": {
                    "200": {
                        "description": "All the avilable categories."
                    },
                    "default": {
                        "description": "something is bad."
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Background": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Standford"
                    },
                    "academic_level": {
                        "type": "string",
                        "example": "University"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Background"
                }
            },
            "Job": {
                "properties": {
                    "company_name": {
                        "type": "string",
                        "example": "Microsoft"
                    },
                    "catg_position_id": {
                        "type": "string",
                        "example": "1"
                    },
                    "experience_years": {
                        "type": "string",
                        "example": "5"
                    },
                    "position_type": {
                        "type": "string",
                        "example": "remote"
                    },
                    "description": {
                        "type": "string",
                        "example": "job description"
                    },
                    "title": {
                        "type": "string",
                        "example": "Back-end lead dev"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Job"
                }
            },
            "Person": {
                "required": [
                    "firstname",
                    "lastname",
                    "email",
                    "gender",
                    "born_date"
                ],
                "properties": {
                    "firstname": {
                        "type": "string",
                        "example": "Jhon"
                    },
                    "lastname": {
                        "type": "string",
                        "example": "Doe"
                    },
                    "email": {
                        "type": "string",
                        "example": "Jhondoe@mail.com"
                    },
                    "gender": {
                        "type": "string",
                        "example": "gender"
                    },
                    "born_date": {
                        "type": "string",
                        "example": "YYYY-MM-DD"
                    },
                    "work_exp_catg": {
                        "type": "string",
                        "example": "1"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Person"
                }
            },
            "WorkExperience": {
                "properties": {
                    "start": {
                        "type": "string",
                        "format": "date",
                        "example": "YYYY-MM-DD"
                    },
                    "end": {
                        "type": "string",
                        "format": "date",
                        "example": "YYYY-MM-DD"
                    },
                    "position": {
                        "type": "string",
                        "example": "Backend lead developer"
                    },
                    "time": {
                        "type": "string",
                        "example": "1"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "WorkExperience"
                }
            }
        }
    }
}